Added TriangleShape

+++ G3D.lib/include/G3D/AABox.h
@@ -130,7 +130,7 @@ public:
     /**
      Distance from corner(0) to the next corner along axis a.
      */
-    inline double extent(int a) const {
+    inline float extent(int a) const {
         debugAssert(a < 3);
         return hi[a] - lo[a];
     }
